pyproject.toml
[tool.poetry]
name = "continuumexcitableelasticity"
version = "0.1.0"
description = "PDE solver for nonlinear odd models"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
py-pde = "^0.17.1"
numpy = "^1.21.4"
matplotlib = "^3.5.0"
llvmlite = "^0.36.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
